+++ lib/Ast.ml
@@ -1371,7 +1371,8 @@ end = struct
        |Pexp_newtype _ | Pexp_open _ | Pexp_override _ | Pexp_pack _
        |Pexp_poly _ | Pexp_record _ | Pexp_send _ | Pexp_sequence _
        |Pexp_setfield _ | Pexp_setinstvar _ | Pexp_tuple _
-       |Pexp_unreachable | Pexp_variant _ | Pexp_while _ | Pexp_hole ->
+       |Pexp_unreachable | Pexp_variant _ | Pexp_while _ | Pexp_hole
+       |Pexp_beginend _ ->
           assert false
       | Pexp_extension (_, ext) -> assert (check_extensions ext)
       | Pexp_object {pcstr_self; pcstr_fields} ->
@@ -1467,7 +1468,8 @@ end = struct
             assert (e1 == exp || e2 == exp)
         | Pexp_extension (_, ext) -> assert (check_extensions ext)
         | Pexp_constant _ | Pexp_ident _ | Pexp_new _ | Pexp_pack _
-         |Pexp_unreachable | Pexp_hole ->
+         |Pexp_unreachable | Pexp_hole
+         |Pexp_beginend None ->
             assert false
         | Pexp_object {pcstr_fields; _} ->
             assert (check_pcstr_fields pcstr_fields)
@@ -1516,6 +1518,7 @@ end = struct
                          true
                      | _ -> e == e ) )
         | Pexp_assert e
+         |Pexp_beginend (Some e)
          |Pexp_constraint (e, _)
          |Pexp_coerce (e, _, _)
          |Pexp_field (e, _)
@@ -2131,7 +2134,9 @@ end = struct
          |Pexp_new _ | Pexp_object _ | Pexp_override _ | Pexp_pack _
          |Pexp_poly _ | Pexp_record _ | Pexp_send _ | Pexp_unreachable
          |Pexp_variant (_, None)
-         |Pexp_hole | Pexp_while _ ->
+         |Pexp_hole | Pexp_while _
+         |Pexp_beginend (Some _)
+         |Pexp_beginend None ->
             false
       in
       Exp.mem_cls cls exp
@@ -2210,7 +2215,9 @@ end = struct
        |Pexp_new _ | Pexp_object _ | Pexp_override _ | Pexp_pack _
        |Pexp_poly _ | Pexp_record _ | Pexp_send _ | Pexp_unreachable
        |Pexp_variant (_, None)
-       |Pexp_hole | Pexp_while _ ->
+       |Pexp_hole | Pexp_while _
+       |Pexp_beginend (Some _)
+       |Pexp_beginend None ->
           false
     in
     Hashtbl.find_or_add marked_parenzed_inner_nested_match exp