+++ CHANGES.md
@@ -16,6 +16,8 @@
 
   + New syntax `(*= ... *)` for verbatim comments (#2028, @gpetiot)
 
+  + Preserve the begin-end construction in the AST (#1785, @gpetiot)
+
 ### 0.21.0 (2022-02-25)
 
 #### Bug fixes